remove `@vue/typescript`
