+++ packages/tsc/package.json
@@ -17,8 +17,8 @@
 		"vue-tsc": "./bin/vue-tsc.js"
 	},
 	"dependencies": {
+		"@volar/typescript": "~1.10.4",
 		"@vue/language-core": "1.8.20",
-		"@vue/typescript": "1.8.20",
 		"semver": "^7.5.4"
 	},
 	"peerDependencies": {
