+++ packages/typescript-plugin/package.json
@@ -13,8 +13,7 @@
 		"directory": "packages/typescript-plugin"
 	},
 	"dependencies": {
-		"@vue/language-core": "1.8.20",
-		"@vue/typescript": "1.8.20",
-		"vscode-uri": "^3.0.8"
+		"@volar/typescript": "~1.10.4",
+		"@vue/language-core": "1.8.20"
 	}
 }
