--- packages/typescript/tsconfig.json
@@ -1,11 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"outDir": "out",
-		"rootDir": "src",
-	},
-	"include": [ "src" ],
-	"references": [
-		{ "path": "../language-core/tsconfig.json" },
-	],
-}
